+++ data-access-tests/Infrastructure/Contexts/TestEfContext.cs
@@ -1,18 +1,19 @@
 ﻿using System;
-using CSGOStats.Infrastructure.DataAccess.Contexts;
+using CSGOStats.Infrastructure.DataAccess.Contexts.EF;
 using CSGOStats.Infrastructure.DataAccess.Tests.Infrastructure.Entities;
 using Microsoft.EntityFrameworkCore;
 
 namespace CSGOStats.Infrastructure.DataAccess.Tests.Infrastructure.Contexts
 {
-    public class TestContext : BaseDataContext
+    public class TestEfContext : BaseDataContext
     {
-        public TestContext()
+        public TestEfContext()
             : base(new PostgreConnectionSettings(
                 host: "127.0.0.1",
                 database: $"Test-{DateTime.UtcNow.Ticks}",
                 username: "postgres",
-                password: "dotFive1"))
+                password: "dotFive1",
+                isAuditEnabled: false))
         {
         }