+++ data-access-tests/Infrastructure/Repositories/TestRepository.cs
@@ -1,4 +1,5 @@
 ﻿using CSGOStats.Infrastructure.DataAccess.Repositories;
+using CSGOStats.Infrastructure.DataAccess.Repositories.EF;
 using CSGOStats.Infrastructure.DataAccess.Tests.Infrastructure.Contexts;
 using CSGOStats.Infrastructure.DataAccess.Tests.Infrastructure.Entities;
 
@@ -7,11 +8,11 @@ namespace CSGOStats.Infrastructure.DataAccess.Tests.Infrastructure.Repositories
     public class TestRepository : EfRepository<TestEntity>
     {
         public TestRepository()
-            : this(new TestContext())
+            : this(new TestEfContext())
         {
         }
 
-        public TestRepository(TestContext context)
+        public TestRepository(TestEfContext context)
             : base(context)
         {
         }