+++ data-access-tests/TestFixture.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using CSGOStats.Infrastructure.DataAccess.Repositories;
+using CSGOStats.Infrastructure.DataAccess.Repositories.Mongo;
 using CSGOStats.Infrastructure.DataAccess.Tests.Infrastructure.Contexts;
 using CSGOStats.Infrastructure.DataAccess.Tests.Infrastructure.Entities;
 using CSGOStats.Infrastructure.DataAccess.Tests.Infrastructure.Repositories;
@@ -8,22 +9,27 @@ namespace CSGOStats.Infrastructure.DataAccess.Tests
 {
     public class TestFixture : IDisposable
     {
-        private readonly TestContext _context = new TestContext();
+        private readonly TestEfContext _efContext = new TestEfContext();
+        private readonly TestMongoContext _mongoContext = new TestMongoContext();
 
-        public IRepository<TestEntity> Repository { get; }
+        public IRepository<TestEntity> EntityRepository { get; }
+
+        public IRepository<TestDocument> DocumentRepository { get; }
 
         public TestFixture()
         {
-            _context.Database.EnsureCreated();
+            _efContext.Database.EnsureCreated();
 
-            Repository = new TestRepository(_context);
+            EntityRepository = new TestRepository(_efContext);
+            DocumentRepository = new GuidKeyMongoRepository<TestDocument>(_mongoContext);
         }
 
         public void Dispose()
         {
-            _context.Database.EnsureDeleted();
+            _efContext.Database.EnsureDeleted();
 
-            _context.Dispose();
+            _efContext.Dispose();
+            _mongoContext.Dispose();
         }
     }
 }