+++ data-access/Contexts/EF/Conversion/ModelBuildingExtensions.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using NodaTime;
+
+namespace CSGOStats.Infrastructure.DataAccess.Contexts.EF.Conversion
+{
+    public static class ModelBuildingExtensions
+    {
+        public static PropertyBuilder<OffsetDateTime> OffsetDateTime(this PropertyBuilder<OffsetDateTime> x) =>
+            x.HasConversion(new OffsetDateTimeConverter());
+    }
+}